crud_api/requirements.txt
alpaca-backtrader-api==0.13.1
alpaca-trade-api==0.52.0
asgiref==3.3.4
awsebcli==3.20.2
backtrader==1.9.76.123
beautifulsoup4==4.9.3
botocore==1.21.0
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.1
capmonster-python==1.3.0
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.3
configparser==5.0.2
crayons==0.4.0
cryptography==3.4.6
cycler==0.10.0
decorator==5.0.9
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.4
django-cors-headers==3.10.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-restframework==0.0.1
django-storages==1.11.1
django-toolbelt==0.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docutils==0.16
ffmpeg==1.4
ffmpy==0.3.0
firebase==3.0.1
firebase-admin==4.5.2
future==0.16.0
gcloud==0.18.3
geocoder==1.38.1
geographiclib==1.50
geopy==2.1.0
gevent==21.1.2
gmplot==1.4.1
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
googlemaps==4.4.5
greenlet==1.0.0
grpcio==1.36.1
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
instaloader==1.2
jmespath==0.10.0
joblib==1.0.1
jwcrypto==0.8
kiwisolver==1.3.1
matplotlib==3.2.2
msgpack==1.0.2
Naked==0.1.31
numpy==1.20.1
oauth2client==4.1.3
opencv-python==4.5.2.54
packaging==20.9
pandas==1.2.3
pathspec==0.5.9
pdfkit==0.6.1
pep8==1.7.1
Pillow==8.1.1
prettyprint==0.1.5
prettytable==2.1.0
proto-plus==1.17.0
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydub==0.25.1
pyparsing==2.4.7
pypiwin32==223
pytesseract==0.3.7
python-barcode==0.13.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==302
PyYAML==5.4.1
ratelim==0.1.6
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
s3transfer==0.3.4
scikit-learn==0.24.1
scipy==1.6.1
selenium==3.141.0
semantic-version==2.8.5
Shapely==1.7.1
shellescape==3.8.1
six==1.14.0
sklearn==0.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
sseclient==0.0.27
static3==0.7.0
termcolor==1.1.0
tesseract==0.1.3
threadpoolctl==2.1.0
toolz==0.11.1
trading-calendars==2.1.1
uritemplate==3.0.1
urllib3==1.26.7
wcwidth==0.1.9
webdriver-manager==3.3.0
websocket-client==0.58.0
websockets==8.1
whitenoise==5.2.0
wkhtmltopdf==0.2
zope.event==4.5.0
zope.interface==5.3.0
